bring back freeze button
