+++ src/App.tsx
@@ -12,6 +12,7 @@ import ReactModal from 'react-modal';
 import MusicButton from './components/buttons/MusicButton.tsx';
 import Button from './components/buttons/Button.tsx';
 import InteractButton from './components/buttons/InteractButton.tsx';
+import FreezeButton from './components/FreezeButton.tsx';
 
 export default function Home() {
   const [helpModalOpen, setHelpModalOpen] = useState(false);
@@ -85,6 +86,7 @@ export default function Home() {
 
         <footer className="absolute bottom-0 left-0 w-full flex items-center mt-4 gap-3 p-6 flex-wrap pointer-events-none">
           <div className="flex gap-4 flex-grow pointer-events-none">
+            <FreezeButton />
             <MusicButton />
             <Button href="https://github.com/get-convex/ai-town" imgUrl={starImg}>
               Star
